Refactor tests to custom render function
