+++ src/context/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import { AuthProvider } from "./auth-context";
+
+const AppProviders = ({ children }) => <AuthProvider>{children}</AuthProvider>;
+
+export default AppProviders;
